requirements.txt
setuptools
numpy
pandas
google-cloud-storage
pyyaml
tensorflow
scikit-learn
comet_ml
dvc